Move enhance functionality to own modules
